+++ src/main/kotlin/com/github/pooryam92/vimcoach/features/tips/domain/VimTip.kt
@@ -5,5 +5,6 @@ data class VimTip(
     var details: List<String> = emptyList(),
     var category: List<String> = emptyList(),
     var config: TipConfig? = null,
-    var mnemonic: String? = null
+    var mnemonic: String? = null,
+    var advanced: Boolean = false
 )